# Speed up SampleWeightComputer