--- a/core/src/prediction/LLCausalPredictionStrategy.cpp
+++ b/core/src/prediction/LLCausalPredictionStrategy.cpp
@@ -39,14 +39,14 @@ size_t LLCausalPredictionStrategy::prediction_length() const {
 
 std::vector<double> LLCausalPredictionStrategy::predict(
         size_t sampleID,
-        const std::unordered_map<size_t, double>& weights_by_sampleID,
+        const Eigen::SparseVector<double>& weights_by_sampleID,
         const Data& train_data,
         const Data& test_data) const {
 
   // Number of predictor variables to use in local linear regression step
   size_t num_variables = linear_correction_variables.size();
 
-  size_t num_nonzero_weights = weights_by_sampleID.size();
+  size_t num_nonzero_weights = weights_by_sampleID.nonZeros();
   size_t num_lambdas = lambdas.size();
 
   // Creating a vector of neighbor weights weights
@@ -60,9 +60,9 @@ std::vector<double> LLCausalPredictionStrategy::predict(
   Eigen::MatrixXd weights_vec = Eigen::VectorXd::Zero(num_nonzero_weights);
   {
     size_t i = 0;
-    for (const auto& it : weights_by_sampleID) {
-      size_t index = it.first;
-      double weight = it.second;
+    for (Eigen::SparseVector<double>::InnerIterator it(weights_by_sampleID); it; ++it) {
+      size_t index = it.index();
+      double weight = it.value();
       indices[i] = index;
       weights_vec(i) = weight;
       i++;
@@ -147,7 +147,7 @@ std::vector<double> LLCausalPredictionStrategy::predict(
 std::vector<double> LLCausalPredictionStrategy::compute_variance(
         size_t sampleID,
         const std::vector<std::vector<size_t>>& samples_by_tree,
-        const std::unordered_map<size_t, double>& weights_by_sampleID,
+        const Eigen::SparseVector<double>& weights_by_sampleID,
         const Data& train_data,
         const Data& test_data,
         size_t ci_group_size) const {
@@ -155,7 +155,7 @@ std::vector<double> LLCausalPredictionStrategy::compute_variance(
   double lambda = lambdas[0];
 
   size_t num_variables = linear_correction_variables.size();
-  size_t num_nonzero_weights = weights_by_sampleID.size();
+  size_t num_nonzero_weights = weights_by_sampleID.nonZeros();
 
   std::vector<size_t> sample_index_map(train_data.get_num_rows());
   std::vector<size_t> indices(num_nonzero_weights);
@@ -163,9 +163,9 @@ std::vector<double> LLCausalPredictionStrategy::compute_variance(
   Eigen::MatrixXd weights_vec = Eigen::VectorXd::Zero(num_nonzero_weights);
   {
     size_t i = 0;
-    for (auto& it : weights_by_sampleID) {
-      size_t index = it.first;
-      double weight = it.second;
+    for (Eigen::SparseVector<double>::InnerIterator it(weights_by_sampleID); it; ++it) {
+      size_t index = it.index();
+      double weight = it.value();
       indices[i] = index;
       sample_index_map[index] = i;
       weights_vec(i) = weight;