--- a/core/src/prediction/QuantilePredictionStrategy.cpp
+++ b/core/src/prediction/QuantilePredictionStrategy.cpp
@@ -34,12 +34,12 @@ size_t QuantilePredictionStrategy::prediction_length() const {
 
 std::vector<double> QuantilePredictionStrategy::predict(
     size_t prediction_sample,
-    const std::unordered_map<size_t, double>& weights_by_sample,
+    const Eigen::SparseVector<double>& weights_by_sample,
     const Data& train_data,
     const Data& data) const {
   std::vector<std::pair<size_t, double>> samples_and_values;
-  for (const auto& entry : weights_by_sample) {
-    size_t sample = entry.first;
+  for (Eigen::SparseVector<double>::InnerIterator it(weights_by_sample); it; ++it) {
+    size_t sample = it.index();
     samples_and_values.emplace_back(sample, train_data.get_outcome(sample));
   }
 
@@ -47,8 +47,9 @@ std::vector<double> QuantilePredictionStrategy::predict(
 }
 
 std::vector<double> QuantilePredictionStrategy::compute_quantile_cutoffs(
-    const std::unordered_map<size_t, double>& weights_by_sample,
+    const Eigen::SparseVector<double>& weights_by_sample,
     std::vector<std::pair<size_t, double>>& samples_and_values) const {
+
   std::sort(samples_and_values.begin(),
             samples_and_values.end(),
             [](std::pair<size_t, double> first_pair, std::pair<size_t, double> second_pair) {
@@ -67,7 +68,7 @@ std::vector<double> QuantilePredictionStrategy::compute_quantile_cutoffs(
     size_t sample = entry.first;
     double value = entry.second;
 
-    cumulative_weight += weights_by_sample.at(sample);
+    cumulative_weight += weights_by_sample.coeff(sample);
     while (quantile_it != quantiles.end() && cumulative_weight >= *quantile_it) {
       quantile_cutoffs.push_back(value);
       ++quantile_it;
@@ -84,7 +85,7 @@ std::vector<double> QuantilePredictionStrategy::compute_quantile_cutoffs(
 std::vector<double> QuantilePredictionStrategy::compute_variance(
     size_t sampleID,
     const std::vector<std::vector<size_t>>& samples_by_tree,
-    const std::unordered_map<size_t, double>& weights_by_sampleID,
+    const Eigen::SparseVector<double>& weights_by_sampleID,
     const Data& train_data,
     const Data& data,
     size_t ci_group_size) const {