--- a/core/src/prediction/QuantilePredictionStrategy.cpp
+++ b/core/src/prediction/QuantilePredictionStrategy.cpp
@@ -49,7 +49,6 @@ std::vector<double> QuantilePredictionStrategy::predict(
 std::vector<double> QuantilePredictionStrategy::compute_quantile_cutoffs(
     const Eigen::SparseVector<double>& weights_by_sample,
     std::vector<std::pair<size_t, double>>& samples_and_values) const {
-
   std::sort(samples_and_values.begin(),
             samples_and_values.end(),
             [](std::pair<size_t, double> first_pair, std::pair<size_t, double> second_pair) {