--- a/core/src/prediction/QuantilePredictionStrategy.h
+++ b/core/src/prediction/QuantilePredictionStrategy.h
@@ -34,20 +34,20 @@ public:
   size_t prediction_length() const;
 
   std::vector<double> predict(size_t prediction_sample,
-    const std::unordered_map<size_t, double>& weights_by_sample,
+    const Eigen::SparseVector<double>& weights_by_sample,
     const Data& train_data,
     const Data& data) const;
 
   std::vector<double> compute_variance(
       size_t sampleID,
       const std::vector<std::vector<size_t>>& samples_by_tree,
-      const std::unordered_map<size_t, double>& weights_by_sampleID,
+      const Eigen::SparseVector<double>& weights_by_sampleID,
       const Data& train_data,
       const Data& data,
       size_t ci_group_size) const;
 
 private:
-  std::vector<double> compute_quantile_cutoffs(const std::unordered_map<size_t, double>& weights_by_sample,
+  std::vector<double> compute_quantile_cutoffs(const Eigen::SparseVector<double>& weights_by_sample,
                                                std::vector<std::pair<size_t, double>>& samples_and_values) const;
 
   std::vector<double> quantiles;