--- a/core/src/prediction/QuantilePredictionStrategy.h
+++ b/core/src/prediction/QuantilePredictionStrategy.h
@@ -48,7 +48,9 @@ public:
 
 private:
   std::vector<double> compute_quantile_cutoffs(const Eigen::SparseVector<double>& weights_by_sample,
-                                               std::vector<std::pair<size_t, double>>& samples_and_values) const;
+                                               std::vector<
+                                               std::pair<size_t, std::pair<double, double>>
+                                               >& samples_and_values) const;
 
   std::vector<double> quantiles;
 };