--- a/core/src/prediction/SurvivalPredictionStrategy.cpp
+++ b/core/src/prediction/SurvivalPredictionStrategy.cpp
@@ -37,7 +37,7 @@ size_t SurvivalPredictionStrategy::prediction_length() const {
 }
 
 std::vector<double> SurvivalPredictionStrategy::predict(size_t prediction_sample,
-    const std::unordered_map<size_t, double>& weights_by_sample,
+    const Eigen::SparseVector<double>& weights_by_sample,
     const Data& train_data,
     const Data& data) const {
   // the event times will always range from 0, ..., num_failures
@@ -45,9 +45,9 @@ std::vector<double> SurvivalPredictionStrategy::predict(size_t prediction_sample
   std::vector<double> count_failure(num_failures + 1);
   std::vector<double> count_censor(num_failures + 1);
   double sum = 0;
-  for (const auto& entry : weights_by_sample) {
-    size_t sample = entry.first;
-    double forest_weight = entry.second;
+  for (Eigen::SparseVector<double>::InnerIterator it(weights_by_sample); it; ++it) {
+    size_t sample = it.index();
+    double forest_weight = it.value();
     size_t failure_time = train_data.get_outcome(sample);
     double sample_weight = train_data.get_weight(sample);
     if (train_data.is_failure(sample)) {
@@ -115,7 +115,7 @@ std::vector<double> SurvivalPredictionStrategy::predict_nelson_aalen(
 std::vector<double> SurvivalPredictionStrategy::compute_variance(
     size_t sample,
     const std::vector<std::vector<size_t>>& samples_by_tree,
-    const std::unordered_map<size_t, double>& weights_by_sampleID,
+    const Eigen::SparseVector<double>& weights_by_sampleID,
     const Data& train_data,
     const Data& data,
     size_t ci_group_size) const {