--- a/core/src/prediction/SurvivalPredictionStrategy.h
+++ b/core/src/prediction/SurvivalPredictionStrategy.h
@@ -50,14 +50,14 @@ public:
   size_t prediction_length() const;
 
   std::vector<double> predict(size_t prediction_sample,
-    const std::unordered_map<size_t, double>& weights_by_sample,
+    const Eigen::SparseVector<double>& weights_by_sample,
     const Data& train_data,
     const Data& data) const;
 
   std::vector<double> compute_variance(
     size_t sample,
     const std::vector<std::vector<size_t>>& samples_by_tree,
-    const std::unordered_map<size_t, double>& weights_by_sampleID,
+    const Eigen::SparseVector<double>& weights_by_sampleID,
     const Data& train_data,
     const Data& data,
     size_t ci_group_size) const;