--- a/core/src/prediction/collector/SampleWeightComputer.h
+++ b/core/src/prediction/collector/SampleWeightComputer.h
@@ -18,25 +18,20 @@
 #ifndef GRF_SAMPLEWEIGHTCOMPUTER_H
 #define GRF_SAMPLEWEIGHTCOMPUTER_H
 
+#include "Eigen/Sparse"
 #include "forest/Forest.h"
 
-#include <unordered_map>
 #include <vector>
 
 namespace grf {
 
 class SampleWeightComputer {
 public:
-  std::unordered_map<size_t, double> compute_weights(size_t sample,
-                                                     const Forest& forest,
-                                                     const std::vector<std::vector<size_t>>& leaf_nodes_by_tree,
-                                                     const std::vector<std::vector<bool>>& valid_trees_by_sample) const;
-
-private:
-  void add_sample_weights(const std::vector<size_t>& samples,
-                          std::unordered_map<size_t, double>& weights_by_sample) const;
-
-  void normalize_sample_weights(std::unordered_map<size_t, double>& weights_by_sample) const;
+  Eigen::SparseVector<double> compute_weights(size_t sample,
+                                              const Forest& forest,
+                                              const std::vector<std::vector<size_t>>& leaf_nodes_by_tree,
+                                              const std::vector<std::vector<bool>>& valid_trees_by_sample,
+                                              const Data& train_data) const;
 };
 
 } // namespace grf