wip (#359 revived): replace hash table with vector and iterate over non-zero entries to get weights.

speed bench for the 4 prediction strategies that relies on the DefaultCollector (survival benefits the most, why: because surival forests sometimes tend to have very large leaves = much more dense neighborhood = many more hash table polls):

survival speed difference
```
set.seed(1234)
n <- 30000
p <- 5
X <- matrix(runif(n * p), n, p)
W <- rbinom(n, 1, 0.5)
Y.max <- 1
failure.time <- pmin(rexp(n) * X[, 1] + W, Y.max)
censor.time <- 2 * runif(n)
Y <- pmin(failure.time, censor.time)
Y=round(Y,1)
D <- as.integer(failure.time <= censor.time)
Y.grid <- sort(unique(Y))
print(length(Y.grid))
# [1] 11

sf.survival<-survival_forest(cbind(X,W), Y, D, num.trees = 500, compute.oob.predictions = FALSE)

# Point predictions
system.time(S1<- predict(sf.survival, failure.times = Y.grid))

# master
# user  system elapsed 
# 545.602  15.628  84.926 

# this
# user  system elapsed
# 80.823   0.523  12.272
```

causal survival forest:
```
system.time(cs.forest <- causal_survival_forest(X, Y, W, D))
# master
# user   system  elapsed
# 2231.792   63.507  308.017

# this
# user  system elapsed 
# 642.242   8.247  94.731 
```

local linear + the rest is not as drastic since its own prediction logic is the major speed impediment
```
set.seed(123)
n <- 20000
p <- 10
X <- matrix(rnorm(n * p), n, p)
Y <- X[, 1] * rnorm(n)
rf = regression_forest(X, Y, compute.oob.predictions = FALSE)

system.time(pp <- predict(rf, linear.correction.variables = 1:p))
# master
# user  system elapsed
# 97.176   4.345  20.989

# this
# user  system elapsed
# 67.802   2.108  15.156
```

local linear causal:
```
n <- 5000
p <- 10
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
cf <- causal_forest(X, Y, W, compute.oob.predictions = FALSE)

X.test <- matrix(rnorm(30000 * p), 30000, p)
system.time(pp <- predict(cf, X.test, linear.correction.variables = 1:p))

# master 
# user  system elapsed 
# 142.339   3.981  20.287 

# this
# user  system elapsed 
# 86.732   2.291  12.879 
```

quantile forest:
```
n <- 5000
p <- 10
X <- matrix(rnorm(n * p), n, p)
Y <- X[, 1] * rnorm(n)
q.forest <- quantile_forest(X, Y, quantiles = c(0.1, 0.5, 0.9), compute.oob.predictions = FALSE)

X.test <- matrix(rnorm(1e5 * p), 1e5, p)
system.time(q.hat <- predict(q.forest, X.test))

# master
# user  system elapsed 
# 157.440   2.471  23.106 

# this
# user  system elapsed 
# 99.107   1.535  14.738 
```

[large n gist](https://gist.github.com/erikcs/f9aad462d718d6a71f50b29ff090f31b)