@erikcs 2021-04-15
Closing for now. 

Part of the reason SampleWeightComputer can be slow is estimation-specific: if you fit a forest with many large nodes the weighting function will be very dense -> very many hash table lookups. For sparse weighting it is not unreasonably slow.

Maybe there are other directions to tackle this from: if there are ways to tune STL's unordered_map to better handle (rare?) dense weight data, i.e. reducing hash collisions by reserving many buckets, etc...